requirements.txt
alembic==1.18.4
attrs==26.1.0
blinker==1.9.0
boto3==1.42.59
botocore==1.42.59
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
cryptography==46.0.5
Deprecated==1.3.1
Faker==40.4.0
flasgger==0.9.7.1
Flask==3.1.2
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Limiter==4.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
git-filter-repo==2.47.0
greenlet==3.3.1
httplib2==0.31.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
limits==5.8.0
Mako==1.3.10
MarkupSafe==3.0.3
mistune==3.2.0
oauthlib==3.3.1
ordered-set==4.1.0
packaging==26.2
proto-plus==1.27.1
protobuf==6.33.5
psycopg2==2.9.11
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
slugify==0.0.1
SQLAlchemy==2.0.46
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
Werkzeug==3.1.5
wrapt==2.1.2
